chore: move examples
The other two pages (Dashboard, Disk Image) are in folders, this just moves
examples to follow the same pattern and reduce files in the root folder.

Favours /@/ imports style for the changes.

Fixes #1293.
