+++ packages/frontend/src/lib/examples/ExampleDetails.spec.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2024 Red Hat, Inc.
+ * Copyright (C) 2024-2025 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ import '@testing-library/jest-dom/vitest';
 import { render, fireEvent, screen } from '@testing-library/svelte';
 import { expect, test, vi } from 'vitest';
 import ExampleDetails from './ExampleDetails.svelte';
-import { bootcClient } from '../api/client';
+import { bootcClient } from '/@/api/client';
 import { router } from 'tinro';
 import type { Example } from '/@shared/src/models/examples';
