+++ packages/frontend/src/lib/examples/ExampleDetailsLayout.svelte
