+++ packages/frontend/src/lib/examples/ExamplesCard.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2024 Red Hat, Inc.
+ * Copyright (C) 2024-2025 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,12 +21,12 @@ import { render, screen } from '@testing-library/svelte';
 import { expect, test, vi } from 'vitest';
 import type { Example, Category } from '/@shared/src/models/examples';
 import ExamplesCard from './ExamplesCard.svelte';
-import { bootcClient } from '../api/client';
+import { bootcClient } from '/@/api/client';
 import { tick } from 'svelte';
 import type { ImageInfo } from '@podman-desktop/api';
 
 // Mock bootcClient methods
-vi.mock('../api/client', () => {
+vi.mock('/@/api/client', () => {
   return {
     bootcClient: {
       listBootcImages: vi.fn(),
